part

+++ runner.ts
@@ -35,8 +35,8 @@ export async function run(source : string, config: any) : Promise<[any, compiler
     importObject.js = { memory: memory };
   }
   const wasmSource = `(module
-    (func $print (import "imports" "imported_func") (param i32))
-    (func $printglobal (import "imports" "print_global_func") (param i32) (param i32))
+    (func $print (import "imports" "imported_func") (param i64))
+    (func $printglobal (import "imports" "print_global_func") (param i64) (param i64))
     (import "js" "memory" (memory 1))
     (func (export "exported_func") ${returnType}
       ${compiled.wasmSource}
